refactor: move the TrendingApi into new datasource module
Move the TrendingApi and TrendingDomain into a new datasource:trending module, removing the need for
two different modules
