+++ app/build.gradle
@@ -64,6 +64,7 @@ dependencies {
     implementation project(":libraries:core")
     implementation project(":libraries:database")
     implementation project(":libraries:datasource:series")
+    implementation project(":libraries:datasource:trending")
     implementation project(":libraries:datasource:user")
     implementation project(":libraries:encryption")
     implementation project(":libraries:kitsu")
@@ -74,7 +75,6 @@ dependencies {
     implementation project(":libraries:kitsu:user")
     implementation project(":libraries:server:auth")
     implementation project(":libraries:server:search")
-    implementation project(":libraries:server:trending")
 
     implementation "androidx.appcompat:appcompat:$appcompat_version"
     implementation "androidx.constraintlayout:constraintlayout:$constraint_layout_version"
