+++ features/discover/src/main/java/com/chesire/nekome/app/discover/DiscoverViewModel.kt
@@ -10,7 +10,7 @@ import com.chesire.nekome.core.Resource
 import com.chesire.nekome.core.extensions.postError
 import com.chesire.nekome.core.extensions.postSuccess
 import com.chesire.nekome.core.flags.AsyncState
-import com.chesire.nekome.trending.api.TrendingApi
+import com.chesire.nekome.datasource.trending.remote.TrendingApi
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
