+++ libraries/datasource/trending/.gitignore
