+++ libraries/datasource/trending/build.gradle
