+++ libraries/datasource/trending/src/main/java/com/chesire/nekome/datasource/trending/remote/TrendingApi.kt
@@ -1,11 +1,13 @@
-package com.chesire.nekome.trending.api
+package com.chesire.nekome.datasource.trending.remote
 
 import com.chesire.nekome.core.Resource
+import com.chesire.nekome.datasource.trending.TrendingDomain
 
 /**
  * Methods relating to getting information about trending topics.
  */
 interface TrendingApi {
+
     /**
      * Retrieves the current trending anime.
      */
