+++ libraries/kitsu/trending/build.gradle
@@ -31,8 +31,8 @@ android {
 
 dependencies {
     implementation project(":libraries:core")
+    implementation project(':libraries:datasource:trending')
     implementation project(':libraries:kitsu')
-    implementation project(':libraries:server:trending')
 
     implementation "com.chesire.lintrules:lint-gradle:$lintrules_version"
     implementation "com.chesire.lintrules:lint-xml:$lintrules_version"
