+++ libraries/server/search/src/main/java/com/chesire/nekome/search/api/SearchApi.kt
@@ -6,6 +6,7 @@ import com.chesire.nekome.core.Resource
  * Methods relating to searching for series.
  */
 interface SearchApi {
+
     /**
      * Search for the anime series [title].
      */
