+++ settings.gradle
@@ -10,16 +10,16 @@ include ':app',
     ':libraries:core',
     ':libraries:database',
     ':libraries:datasource:series',
+    ':libraries:datasource:trending',
     ':libraries:datasource:user',
     ':libraries:encryption',
     ':libraries:kitsu',
     ':libraries:kitsu:auth',
     ':libraries:kitsu:library',
     ':libraries:kitsu:search',
-    ':libraries:kitsu:user',
     ':libraries:kitsu:trending',
+    ':libraries:kitsu:user',
     ':libraries:server:auth',
     ':libraries:server:search',
-    ':libraries:server:trending',
 
     ':testing'
